pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jeedoli <ljhx6787@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-ninja = "^1.3.0"
django-allauth = "^65.3.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
cryptography = "^44.0.0"
gunicorn = "^23.0.0"
google-auth = "^2.36.0"
boto3 = "^1.35.86"
django-storages = "^1.14.4"
django-cleanup = "^9.0.0"
django-redis = "^5.4.0"
redis = "^5.2.1"
channels = "^4.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"
asgiref = "^3.8.1"
async-timeout = "^5.0.1"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
uvicorn-worker = "^0.3.0"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.4"}

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pre-commit = "^4.0.1"
django-extensions = "^3.2.3"
ipython = "^8.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
